safari-600.7-branch/LayoutTests/fast/forms/image/image-setrangetext.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/common-setrangetext.js"></script>
</head>
<body>
<script>
description("Test setRangeText() method is not available in image inputs.");
runTestsShouldFail("input", { type: "image" });
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
