safari-600.7-branch/LayoutTests/fast/forms/image/image-setrangetext-expected.txt

Test setRangeText() method is not available in image inputs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Running tests on input with attributes: {"type":"image"}

element.value = '0123456789XYZ'
PASS element.setRangeText('ABC', 0, 0) threw exception Error: InvalidStateError: DOM Exception 11.
PASS element.value is "0123456789XYZ"
PASS successfullyParsed is true

TEST COMPLETE
